// src/form/nxa-form-textarea.d.ts
import { LitElement } from 'lit';
export declare class NxaFormTextarea extends LitElement {
    static styles: import("lit").CSSResult;
    label: string;
    placeholder: string;
    value: string;
    name: string;
    size: 'sm' | 'md' | 'lg';
    required: boolean;
    disabled: boolean;
    readonly: boolean;
    floating: boolean;
    helperText: string;
    rows: number;
    maxRows: number;
    private touched;
    private valid;
    private invalid;
    private isEmpty;
    private getTextareaClasses;
    private get _uniqueId();
    private _handleInput;
    private _adjustHeight;
    private _handleBlur;
    validateOnSubmit(): boolean;
    private _validate;
    firstUpdated(): void;
    updated(changedProperties: Map<string, any>): void;
    render(): import("lit-html").TemplateResult<1>;
}